Implement basic login functionality
This commit adds version checking to the data to ensure weird bugs with
opening a vault from different versions of byteman don't occur. This commit
also adds username login functionality which works by checking for a
file called <user_name>.vault.

+++ include/utils.h
@@ -58,7 +58,7 @@ int FileExists(const char *name);
  * @return 1 on success 
  * @return 0 on failure (EOF or input error)
  */
-int ReadInput(char *buf, const int BUFFER_SIZE);
+int ReadInput(char *buf, const size_t BUFFER_SIZE);
 
 
 /**
